Move diffmap_widget into pyFAI.gui
