+++ pyFAI/gui/setup.py
@@ -32,6 +32,7 @@ from numpy.distutils.misc_util import Configuration
 def configuration(parent_package='', top_path=None):
     config = Configuration('gui', parent_package, top_path)
     config.add_subpackage('test')
+    config.add_subpackage('utils')
     config.add_subpackage('calibration')
     return config
