+++ pyFAI/gui/utils/__init__.py
@@ -34,11 +34,11 @@ __author__ = "Jerome Kieffer"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "20/02/2018"
+__date__ = "05/03/2018"
 __status__ = "production"
 
 
-from . import matplotlib
+from .. import matplotlib
 from silx.gui import qt
 
 main_loop = False
